PD-TWEAK: tweak the logs again......................

+++ code/gnomehw/build.sh
@@ -3,31 +3,35 @@
 set -e
 set -x
 
-rm -rf cmake-build-debug
-mkdir cmake-build-debug
-git checkout cmake-build-debug/.gitignore
-pushd cmake-build-debug
+if [[ $# -eq $# ]]; then
+    rm -rf cmake-build-debug
+    mkdir cmake-build-debug
+    git checkout cmake-build-debug/.gitignore
+    pushd cmake-build-debug
 
-echo "-------------------- CONAN SETUP/BUILD -----------------------" > /dev/null
-conan install ..
-echo "-------------------- CMAKE SETUP -----------------------" > /dev/null
-cmake .. -DCMAKE_BUILD_TYPE=Debug
-echo "-------------------- CMAKE BUILD -----------------------" > /dev/null
-cmake --build .
-./bin/gnomehw
-popd
+    echo "-------------------- CONAN SETUP/BUILD -----------------------" > /dev/null
+    conan install ..
+    echo "-------------------- CMAKE SETUP -----------------------" > /dev/null
+    cmake .. -DCMAKE_BUILD_TYPE=Debug
+    echo "-------------------- CMAKE BUILD -----------------------" > /dev/null
+    cmake --build .
+    ./bin/gnomehw
+    popd
+fi
 
-rm -rf cmake-build-release
-mkdir cmake-build-release
-git checkout cmake-build-debug/.gitignore
-pushd cmake-build-release
+if [[ $# -gt 0 ]]; then
+    rm -rf cmake-build-release
+    mkdir cmake-build-release
+    git checkout cmake-build-debug/.gitignore
+    pushd cmake-build-release
 
-echo "-------------------- CONAN SETUP/BUILD -----------------------" > /dev/null
-conan install ..
-echo "-------------------- CMAKE SETUP -----------------------" > /dev/null
-cmake .. -DCMAKE_BUILD_TYPE=Release
-echo "-------------------- CMAKE BUILD -----------------------" > /dev/null
-cmake --build .
+    echo "-------------------- CONAN SETUP/BUILD -----------------------" > /dev/null
+    conan install ..
+    echo "-------------------- CMAKE SETUP -----------------------" > /dev/null
+    cmake .. -DCMAKE_BUILD_TYPE=Release
+    echo "-------------------- CMAKE BUILD -----------------------" > /dev/null
+    cmake --build .
 
-echo "-------------------- RUN" > /dev/null
-./bin/gnomehw
+    echo "-------------------- RUN" > /dev/null
+    ./bin/gnomehw
+fi
